move samples table into components
