+++ src/components/SamplesTable/ProcessingInformation/ProcessorVersion.scss
