+++ src/components/SamplesTable/ProcessingInformationCell.js
