+++ src/components/SamplesTable/SampleDetailsLoader.js
