+++ src/components/SamplesTable/SamplesTable.scss
